pyproject.toml
[tool.poetry]
name = "crypto-monkas-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "crypto_monkas_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
Flask-RESTful = "^0.3.9"
black = "^22.10.0"
numpy = "^1.23.3"
sympy = "^1.11.1"
requests = "^2.28.1"
more-itertools = "^8.14.0"
opencv-python-headless = "^4.6.0.66"
flask-cors = "^3.0.10"
pillow = "^9.3.0"
pycrypto = "^2.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
